+++ deltakit-explorer/src/deltakit_explorer/__init__.py
@@ -3,7 +3,7 @@ import importlib.metadata
 
 from deltakit_explorer import analysis, visualisation
 from deltakit_explorer._api._client import Client
-from deltakit_explorer._utils._logging import Logging
+from deltakit_core.api._logging import Logging
 
 __version__ = importlib.metadata.version(__package__)
