+++ deltakit-explorer/src/deltakit_explorer/_api/_api_v2_client.py
@@ -16,7 +16,7 @@ if TYPE_CHECKING:
 from deltakit_explorer._api._api_client import APIClient, APIEndpoints
 from deltakit_explorer._api._auth import (get_token,
                                           set_token)
-from deltakit_explorer._utils._logging import Logging
+from deltakit_core.api._logging import Logging
 from deltakit_core.api.enums import DataFormat
 from deltakit_core.types._exceptions import ServerException
 from deltakit_core.types._experiment_types import QECExperimentDefinition
