+++ deltakit-explorer/src/deltakit_explorer/_api/_client.py
@@ -20,7 +20,7 @@ from deltakit_core.api import environment as utils
 from deltakit_explorer._utils._decorators import (
     validate_and_split_decoding, validate_and_split_simulation,
     validate_generation)
-from deltakit_explorer._utils._logging import Logging
+from deltakit_core.api._logging import Logging
 from deltakit_explorer.data._data_analysis import has_leakage
 from deltakit_core.types._experiment_types import (QECExperiment,
                                                        QECExperimentDefinition)
